==> 00_cleaning_and_export.ipynb <==
import pandas as pd
import numpy as np
import re

df = pd.read_csv('reviews.csv', encoding = 'iso-8859-1')
df.head()

df.dtypes

df.iloc[2, 7], df.iloc[2, 8], df.iloc[3, 7], df.iloc[3, 7]

df['Distance'] = df['Distance'].apply(lambda x: re.sub('[\x97]', '', x))
df['Distance'] = df['Distance'].apply(lambda x: re.sub('[\x97]', '', x))
df['Cost'] = df['Cost'].apply(lambda x: re.sub('[\x97]', '', x))
df['Per.Mile'] = df['Per.Mile'].apply(lambda x: re.sub('[\x97]', '', x))
df['Sqft.Moved'] = df['Sqft.Moved'].apply(lambda x: re.sub('[\x97]', '', x))
df.head()

df.iloc[2, 7], df.iloc[2, 8], df.iloc[3, 7], df.iloc[3, 7]

np.sum(df.isnull())

print(df[df['City'].isnull()]['City'])
df.iloc[2225, 2]= 'New York'
df.iloc[2225,]

np.sum(df.isnull())

df.head()

df.rename(columns={'Per.Mile': 'Cost.per.Mile'}, inplace=True)
df['Date'] = pd.to_datetime(df['Date'])
df['Distance'] = df['Distance'].apply(lambda x: float(''.join(x.split()[0].split(','))) if x != '' else np.nan)
df['Cost'] = df['Cost'].apply(lambda x: float(''.join(x.strip('$').split(','))) if x != '' else np.nan)
df['Cost.per.Mile'] = df['Cost.per.Mile'].apply(lambda x: float(''.join(x.strip('$').split(','))) if x != '' else np.nan)
df['Sqft.Moved'] = df['Sqft.Moved'].apply(lambda x: float(''.join(x.split(','))) if x != '' else np.nan)
df.head()

df.dtypes

np.sum(df.isnull())

df[df.duplicated(keep=False)]

df.shape

df = df.drop_duplicates()

df.shape

df.to_csv('reviews_cleaned.csv', index = False)

df2 = pd.read_csv('reviews_cleaned.csv')
df2

df2.shape

df2.dtypes

np.sum(df2.isnull())
