==> 02_bubble_map.ipynb <==
import pandas as pd

df = pd.read_csv('reviews_cleaned.csv')
df.head()

df = df[['Company', 'City', 'State', 'Distance', 'Sqft.Moved']]

df.sample(10)

df['City'] = df['City'].apply(lambda x: ' '.join(x.title() for x in x.split()))

import re
df['City'] = df['City'].apply(lambda x: re.sub('[^\w\s]', '', x))

df.shape

df = df.drop_duplicates()
df.shape

df_geocoder = pd.read_csv('geocoder.csv')
df_geocoder.head()

df_map = pd.merge(df, df_geocoder, how = 'left', on = ['City', 'State'])
df_map.head()

df_map.tail()

df_map.sample(5)

import numpy as np

np.sum(df_map.isnull(), axis = 0)

df_map.shape

df_map = df_map[df_map['latitude'].notnull()]
df_map.shape

import plotly.plotly as py

df_map_count = df_map.groupby(['City', 'State', 'latitude', 'longitude']).size().reset_index()
df_map_count.columns = ['City', 'State', 'latitude', 'longitude', 'count']
df_map_count.head(5)

df_map_count['text'] = df_map_count['City'] + ', ' + df_map_count['State']
df_map_count.head(5)

cities = []
city = dict(
    type = 'scattergeo',
    locationmode = 'USA-states',
    lat = df_map_count['latitude'], 
    lon = df_map_count['longitude'], 
    text = df_map_count['text'],
    marker = dict(
        size = df_map_count['count'],
        color = 'rgb(0,116,217)',
        line = dict(width=0.5, color='rgb(40,40,40)'),
        sizemode = 'area'),
    name = '{0} - {1}'.format(0, df_map_count.shape[1])
)
cities.append(city)

layout = dict(
    title = 'All Reviewer Locations',
    showlegend = True,
    geo = dict(
        scope='usa',
        projection=dict(type='albers usa'),
        showland = True,
        landcolor = 'rgb(217, 217, 217)',
        subunitwidth=1,
        countrywidth=1,
        subunitcolor="rgb(255, 255, 255)",
        countrycolor="rgb(255, 255, 255)"
    ),
)

# plotly username: plotlylearner, gmail
# password: 8+eng+$*
# regenerate the key: https://plot.ly/settings/api#/
import plotly
plotly.tools.set_credentials_file(username='plotlylearner', api_key = 'Or2tF6uyx8YdAKkRR0nY' )

fig = dict(data = cities, layout = layout)
py.iplot(fig, validate=False, filename='d3-bubble-map-all-reviews')

df_map_campany = df_map.groupby(['Company', 'City', 'State', 'latitude', 'longitude']).size().reset_index()
df_map_campany.columns = ['Company', 'City', 'State', 'latitude', 'longitude', 'count']
df_map_campany.head(5)

df_map_campany['text'] = df_map_campany['City'] + ', ' + df_map_campany['State']
df_map_campany.head(5)

df_map_campany.shape

df_map_campany[df_map_campany['Company'] == 'Budget']['Company'].count()

df_map_campany[df_map_campany['Company'] == 'Enterprise']['Company'].count()

df_map_campany[df_map_campany['Company'] == 'Penske']['Company'].count()

df_map_campany[df_map_campany['Company'] == 'Ryder']['Company'].count()

df_map_campany[df_map_campany['Company'] == 'U-Haul']['Company'].count()

df_map_campany.shape

print(df_map_campany.iloc[664,0], df_map_campany.iloc[665,0])

print(df_map_campany.iloc[681,0], df_map_campany.iloc[682,0])

print(df_map_campany.iloc[1565,0], df_map_campany.iloc[1566,0])

print(df_map_campany.iloc[1581,0], df_map_campany.iloc[1582,0])

df_map_campany.iloc[3048,0]

# company: budget, enterprise, penske, ryder, uhaul
limits = [(0,665),(665,682),(682,1566),(1566,1582),(1582,3049)]
colors = ["rgb(64,66,122)","rgb(29,121,81)","rgb(206,174,57)","rgb(166,24,42)","rgb(247,87,63)"]

cities = []
for i in range(len(limits)):
    lim = limits[i]
    df_sub = df_map_campany[lim[0]:lim[1]]
    city = dict(
        type = 'scattergeo',
        locationmode = 'USA-states',
        lat = df_map_campany['latitude'], 
        lon = df_map_campany['longitude'], 
        text = df_map_campany['text'],
        marker = dict(
            size = df_map_campany['count']*8,
            color = colors[i],
            line = dict(width=0.5, color='rgb(40,40,40)'),
            sizemode = 'area'
        ),
        name = '{0} - {1}'.format(lim[0],lim[1])
    )
    cities.append(city)

layout = dict(
    title = 'Company Distribution',
    showlegend = True,
    geo = dict(
        scope='usa',
        projection=dict(type='albers usa'),
        showland = True,
        landcolor = 'rgb(217, 217, 217)',
        subunitwidth=1,
        countrywidth=1,
        subunitcolor="rgb(255, 255, 255)",
        countrycolor="rgb(255, 255, 255)"
    ),
)

fig = dict(data = cities, layout = layout)
py.iplot(fig, validate=False, filename='d3-bubble-map-companies')

# Budget

# company: budget, enterprise, penske, ryder, uhaul
# Budget, Enterprise, Penske, Ryder, U-Haul
# limits = [(0,665),(665,682),(682,1566),(1566,1582),(1582,3049)]
# colors = ["rgb(64,66,122)","rgb(29,121,81)","rgb(206,174,57)","rgb(166,24,42)","rgb(247,87,63)"]

df_sub = df_map_campany[0:665]
df_sub.shape

df_sub.sample(5)

cities = []
city = dict(
    type = 'scattergeo',
    locationmode = 'USA-states',
    lat = df_sub['latitude'], 
    lon = df_sub['longitude'], 
    text = df_sub['text'],
    marker = dict(
        size = df_sub['count']*8,
        color = 'rgb(64,66,122)',
        line = dict(width=0.5, color='rgb(40,40,40)'),
        sizemode = 'area'),
    name = '{0} - {1}'.format(0, df_sub.shape[1])
)
cities.append(city)

layout = dict(
    title = 'Budget Locations',
    showlegend = True,
    geo = dict(
        scope='usa',
        projection=dict(type='albers usa'),
        showland = True,
        landcolor = 'rgb(217,217,217)',
        subunitwidth=1,
        countrywidth=1,
        subunitcolor="rgb(255, 255, 255)",
        countrycolor="rgb(255, 255, 255)"
    ),
)

fig = dict(data = cities, layout = layout)
py.iplot(fig, validate=False, filename='d3-bubble-map-budget')

# Enterprise

# company: budget, enterprise, penske, ryder, uhaul
# Budget, Enterprise, Penske, Ryder, U-Haul
# limits = [(0,665),(665,682),(682,1566),(1566,1582),(1582,3049)]
# colors = ["rgb(64,66,122)","rgb(29,121,81)","rgb(206,174,57)","rgb(166,24,42)","rgb(247,87,63)"]

df_sub = df_map_campany[665:682]
df_sub.shape

df_sub.sample(5)

cities = []
city = dict(
    type = 'scattergeo',
    locationmode = 'USA-states',
    lat = df_sub['latitude'], 
    lon = df_sub['longitude'], 
    text = df_sub['text'],
    marker = dict(
        size = df_sub['count']*30,
        color = 'rgb(29,121,81)',
        line = dict(width=0.5, color='rgb(40,40,40)'),
        sizemode = 'area'),
    name = '{0} - {1}'.format(0, df_sub.shape[1])
)
cities.append(city)

layout = dict(
    title = 'Enterprise Locations',
    showlegend = True,
    geo = dict(
        scope='usa',
        projection=dict(type='albers usa'),
        showland = True,
        landcolor = 'rgb(217,217,217)',
        subunitwidth=1,
        countrywidth=1,
        subunitcolor="rgb(255, 255, 255)",
        countrycolor="rgb(255, 255, 255)"
    ),
)

fig = dict(data = cities, layout = layout)
py.iplot(fig, validate=False, filename='d3-bubble-map-enterprise')

# Penske

# company: budget, enterprise, penske, ryder, uhaul
# Budget, Enterprise, Penske, Ryder, U-Haul
# limits = [(0,665),(665,682),(682,1566),(1566,1582),(1582,3049)]
# colors = ["rgb(64,66,122)","rgb(29,121,81)","rgb(206,174,57)","rgb(166,24,42)","rgb(247,87,63)"]

df_sub = df_map_campany[682:1566]
df_sub.shape

df_sub.sample(5)

cities = []
city = dict(
    type = 'scattergeo',
    locationmode = 'USA-states',
    lat = df_sub['latitude'], 
    lon = df_sub['longitude'], 
    text = df_sub['text'],
    marker = dict(
        size = df_sub['count']*8,
        color = 'rgb(206,174,57)',
        line = dict(width=0.5, color='rgb(40,40,40)'),
        sizemode = 'area'),
    name = '{0} - {1}'.format(0, df_sub.shape[1])
)
cities.append(city)

layout = dict(
    title = 'Penske Locations',
    showlegend = True,
    geo = dict(
        scope='usa',
        projection=dict(type='albers usa'),
        showland = True,
        landcolor = 'rgb(217,217,217)',
        subunitwidth=1,
        countrywidth=1,
        subunitcolor="rgb(255, 255, 255)",
        countrycolor="rgb(255, 255, 255)"
    ),
)

fig = dict(data = cities, layout = layout)
py.iplot(fig, validate=False, filename='d3-bubble-map-penske')

# Ryder

# company: budget, enterprise, penske, ryder, uhaul
# Budget, Enterprise, Penske, Ryder, U-Haul
# limits = [(0,665),(665,682),(682,1566),(1566,1582),(1582,3049)]
# colors = ["rgb(64,66,122)","rgb(29,121,81)","rgb(206,174,57)","rgb(166,24,42)","rgb(247,87,63)"]

df_sub = df_map_campany[1566:1582]
df_sub.shape

df_sub.sample(5)

cities = []
city = dict(
    type = 'scattergeo',
    locationmode = 'USA-states',
    lat = df_sub['latitude'], 
    lon = df_sub['longitude'], 
    text = df_sub['text'],
    marker = dict(
        size = df_sub['count']*30,
        color = 'rgb(166,24,42)',
        line = dict(width=0.5, color='rgb(40,40,40)'),
        sizemode = 'area'),
    name = '{0} - {1}'.format(0, df_sub.shape[1])
)
cities.append(city)

layout = dict(
    title = 'Ryder Locations',
    showlegend = True,
    geo = dict(
        scope='usa',
        projection=dict(type='albers usa'),
        showland = True,
        landcolor = 'rgb(217,217,217)',
        subunitwidth=1,
        countrywidth=1,
        subunitcolor="rgb(255, 255, 255)",
        countrycolor="rgb(255, 255, 255)"
    ),
)

fig = dict(data = cities, layout = layout)
py.iplot(fig, validate=False, filename='d3-bubble-map-ryder')

# U-Haul

# company: budget, enterprise, penske, ryder, uhaul
# Budget, Enterprise, Penske, Ryder, U-Haul
# limits = [(0,665),(665,682),(682,1566),(1566,1582),(1582,3049)]
# colors = ["rgb(64,66,122)","rgb(29,121,81)","rgb(206,174,57)","rgb(166,24,42)","rgb(247,87,63)"]

df_sub = df_map_campany[1582:]
df_sub.shape

df_sub.sample(5)

cities = []
city = dict(
    type = 'scattergeo',
    locationmode = 'USA-states',
    lat = df_sub['latitude'], 
    lon = df_sub['longitude'], 
    text = df_sub['text'],
    marker = dict(
        size = df_sub['count']*8,
        color = 'rgb(247,87,63)',
        line = dict(width=0.5, color='rgb(40,40,40)'),
        sizemode = 'area'),
    name = '{0} - {1}'.format(0, df_sub.shape[1])
)
cities.append(city)

layout = dict(
    title = 'U-Haul Locations',
    showlegend = True,
    geo = dict(
        scope='usa',
        projection=dict(type='albers usa'),
        showland = True,
        landcolor = 'rgb(217,217,217)',
        subunitwidth=1,
        countrywidth=1,
        subunitcolor="rgb(255, 255, 255)",
        countrycolor="rgb(255, 255, 255)"
    ),
)

fig = dict(data = cities, layout = layout)
py.iplot(fig, validate=False, filename='d3-bubble-map-uhaul')

df_geo = df[['City', 'State']]
df_geo.head()

df_geo.shape

df_geo = df_geo.drop_duplicates()

df_geo.shape

from geopy import geocoders
from geopy.geocoders import Nominatim
# geopy username: geopylearner, gmail
# geopy password: 8+eng+$*

geolocator = Nominatim(user_agent="geopylearner")

location = geolocator.geocode("New York, NY")
print(location.latitude, location.longitude)
location = geolocator.geocode("Houston, TX")
print(location.latitude, location.longitude)

type(location)

type(location.longitude)

a, b = location.latitude, location.longitude
a, b

location = geolocator.geocode('Ortley Beach Toms River Twpshp, NJ')
# print(location.latitude, location.longitude)
location == None

import numpy as np

from geopy.exc import GeocoderTimedOut

df_geo = df_geo.sort_values(['State', 'City'])
df_geo['latitude'] = 0.0
df_geo['longitude'] = 0.0
df_geo = df_geo.reset_index(drop=True)
df_geo.head(5)

df_geo.iloc[1:3]

def do_geocode(address):
    try:
        geolocator = Nominatim(user_agent="geopylearner")
        return geolocator.geocode(address)
    except GeocoderTimedOut as e:
        return do_geocode(address)

def find_geocoder_func(x):
    addr = x['City'] + ', ' + x['State']
    print(addr)
    location = do_geocode(addr)
    if location == None:
        x['latitude'] = np.nan
        x['longitude'] = np.nan
    else:
        x['latitude'] = location.latitude
        x['longitude'] = location.longitude
    return x

df_geo.iloc[2240:2246] = df_geo.iloc[2240:2246].apply(find_geocoder_func, axis = 1)

df_geo.iloc[2238:2242]

df_geo.tail(10)

A: AL AK AZ AR
C: CA CO CT
D: DE DC
F: FL
G: GA
H: HI
I: ID IL IN IA
K: KS KY
L: LA
M: ME MD MA MI MN MS MO MT
N: NE NV NH NJ NM NY NC ND
O: OH OK OR
P: PA
R: RI
S: SC SD
T: TN TX
U: UT
V: VT VA
W: WA WV WI WY

IPv4 Address. . . . . . . . . . . : 192.168.0.104
<br> Subnet Mask . . . . . . . . . . . : 255.255.255.0
<br> Default Gateway . . . . . . . . . : 192.168.0.1

df_geo[df_geo['State'].str.startswith('A')]['State'].count()

df_geo[df_geo['State'].str.startswith('C')]['State'].count()

df_geo[df_geo['State'].str.startswith('D')]['State'].count()

df_geo[df_geo['State'].str.startswith('F')]['State'].count()

df_geo[df_geo['State'].str.startswith('G')]['State'].count()

df_geo[df_geo['State'].str.startswith('I')]['State'].count()

df_geo[df_geo['State'].str.startswith('K')]['State'].count()

df_geo[df_geo['State'].str.startswith('L')]['State'].count()

df_geo[df_geo['State'].str.startswith('M')]['State'].count()

df_geo[df_geo['State'].str.startswith('N')]['State'].count()

df_geo[df_geo['State'].str.startswith('O')]['State'].count()

df_geo[df_geo['State'].str.startswith('P')]['State'].count()

df_geo[df_geo['State'].str.startswith('R')]['State'].count()

df_geo[df_geo['State'].str.startswith('S')]['State'].count()

df_geo[df_geo['State'].str.startswith('T')]['State'].count()

df_geo[df_geo['State'].str.startswith('U')]['State'].count()

df_geo[df_geo['State'].str.startswith('V')]['State'].count()

df_geo[df_geo['State'].str.startswith('W')]['State'].count()

df_geo = df_geo.apply(find_geocoder_func, axis = 1)

df_geo['City'] = df_geo['City'].apply(lambda x: ' '.join(x.title() for x in x.split()))

df_geo.shape

df_geo = df_geo.drop_duplicates()

df_geo.shape

df_geo = df_geo.drop_duplicates()
df_geo.shape

df_geo.to_csv('geocoder.csv', index = False)
